requirements.txt
streamlit==1.44.1
fastapi==0.115.12
uvicorn==0.34.2
langchain==0.3.25
langchain-openai==0.3.18
langchain-community==0.3.24
langchain-chroma==0.2.4
langchain-core==0.3.83
chromadb==0.6.3
openai==1.78.1
python-dotenv==1.1.0
pandas==2.2.3
tiktoken==0.9.0
httpx==0.28.1
pydantic==2.11.3
requests==2.32.3